Added RepositoryBase implementation for checking if an object is present (id: int -> long)
- Updated Repository Interface to use Long instead of int for the identifier column

+++ src/main/java/com/training/java/transaction/tracker/repository/Repository.java
@@ -7,15 +7,15 @@ import java.util.List;
 public interface Repository<T> {
   T create(T object) throws Exception;
 
-  boolean remove(int id) throws SQLException;
+  boolean remove(long id) throws SQLException;
 
   void update(T object) throws SQLException;
 
   List<T> fetchAll() throws SQLException;
 
-  boolean doesIdExist(int id) throws SQLException;
+  boolean doesIdExist(long id) throws SQLException;
 
-  T fetchById(int id) throws SQLException;
+  T fetchById(long id) throws SQLException;
 
   boolean isValid(T object);
 }
